Deprecating some old code and consolidating other code. Starting to move class definitions for datasets & dataset configs out of make_catalogs.py
