+++ lightning_hydra_classifiers/data/datamodules.py
@@ -15,7 +15,7 @@ from omegaconf import DictConfig
 from contrastive_learning.data.pytorch.pnas import PNASLightningDataModule
 from contrastive_learning.data.pytorch.extant import ExtantLightningDataModule
 from contrastive_learning.data.pytorch.fossil import FossilLightningDataModule
-from contrastive_learning.data.pytorch.common import LeavesLightningDataModule
+# from contrastive_learning.data.pytorch.common import LeavesLightningDataModule
 from contrastive_learning.data.pytorch.utils.file_utils import ensure_dir_exists
 __all__ = ["get_datamodule", "fetch_datamodule_from_dataset_artifact"]
 
@@ -75,7 +75,7 @@ def get_datamodule(data_config: DictConfig):
 
 def fetch_datamodule_from_dataset_artifact(dataset_config: Box, 
                                            artifact_config: Box,
-                                           run_or_api=None) -> Tuple[LeavesLightningDataModule, wandb.Artifact]:
+                                           run_or_api=None) -> Tuple["LeavesLightningDataModule", wandb.Artifact]:
     """
     Download a versioned dataset artifact from wandb and output a configured datamodule.
