+++ lightning_hydra_classifiers/utils/__init__.py
@@ -2,8 +2,9 @@ from . import registry
 
 from . import common_utils
 from . import plot_utils
+from . import file_utils
+from . import template_utils
 from . import ResizeRight
-from . import dataset_management_utils, file_utils, inference_example, model_utils, template_utils, metric_utils
-
+from . import dataset_management_utils, model_utils, metric_utils
 from . import logging_utils
